Add files via upload
This is the application by using a single web page we can login and can answer the questions.On submiting the answers we can view the questions for which we gave the correct anwers.There is a countdown timer included.It alerts when the times up. Can Easily understandable code.